Clean up and improve album library screen
